+++ src/providers/index.ts
@@ -1,7 +1,5 @@
 export {
     DevXToolsProvider,
-    AESSettingsPanel,
     ConfigSidebarPanel,
     getGeneralPreferences,
-    getAesKeyIdentifiers,
 } from './devxToolsProvider';